Add links, fix broken link, and minor edits.
Add links for the external non-Python dependencies.  Make it easier to locate and install these dependencies.
The 'legacy' link in the application structure document is incorrect.
Minor typographical errors.
The description of the package_roles() xmlrpc method is incorrect.  It returns a list of [role_name, user] instead of tuples:
>>> import xmlrpclib
>>> client = xmlrpclib.ServerProxy('https://pypi.python.org/pypi')
>>> client.package_roles('warehouse')
[['Owner', 'dstufft'], ['Maintainer', 'python-infra'], ['Owner', 'python-infra']]