Rename models to db
The "models" moniker wasn't very accurate and mostly was left over
from a time when I was going to make these more ORM like. What these
really are is just classes that contain simple methods for hitting
the database (or another datastore) so we'll just call them db to
be more accurate.