Use Camo even in development
Development should match production as much as possible, so we'll
run camo even in development. However we don't know what the domain
of the local site is development so we need a small hack to support
late binding of the actual hostname (but not port number).

This also means that camo is no longer an optional part of a
deployment.