Configure logging with a per request logger
Use structlog to output structured JSON based logging and include
a shim to enable stdlib based logging to still be rendered
similarly to a structlog based logger.

In addition we'll generate a unique id per request and set up a
logger on the request object which enables logging inside of that
request which includes the request id.

This does not include shuffling the Gunicorn logs through a JSON
like output.