Default transactions to use the SERIALIZABLE isolation level
A number of our views expect to be able to do multiple queries from
the database without the state changing between views. This is
exactly what the SERIALIZABLE transaction isolation level exists
to solve. However we can't easily set the transaction isolation for
a single view, only for all views. Given that, we'll set that
isolation and configure pyramid_tm so that it can retry views which
have failed with a retryable exception.