Implement a read only route predicate
By abusing the route predicate system we can create a predicate
which will mark a route as read only. We can then use this to
create a database session which is serializable, read only, and
deferrable. We need this for any view which adds the
X-PyPI-Last-Serial header.

There are some problems with this approach:

* It's an abuse of route predicates, which is less than ideal.
* You can't tell that a view is read only without looking at the
  route.
* We have a second engine (but the pools are shared) which
  implements the read-only portions.