Implement a translation system which defers translation
We want to be able to translate values, but not rely on thread
local storage of the current transaltion or force a manual
translation API.

By using the Jinja2 finalize function we can delay translation
until right before Jinja2 actually renders a template. This allows
us to override the way Jinja2 renders the TranslationString and
call the underlying gettext API transparently.