Don't restore the session after making it invalid
The architecture of Pyramid is such that rendering a response
happens outside of the view callable being called. This means that
a template could access request.user and it would work because the
template rendered after request.session was restored.

To fix this, we'll just not restore the session back again after it
has been replaced with an invalid one.