Rearchitect file serving to better utilize S3
Originally we used the pyfilesystem library to implement file
storage. However when migrating PyPI to S3 it was determined that
this library is extremely inefficient when it comes to what API
calls it makes to S3. Instead we'll utilize boto3 directly and
create our own storage classes to act as an abstraction for where
files are being stored.