Long filenames break rendering.
Fixes #338.

I opted to fix this with a CSS overflow property. In most cases, it
shouldn't look different, but for very long file names, such as in the
Pillow project, you'll be able to see it.

I notice that the compiled CSS is under version control; that seems
undesirable to me. Would there be any objections to adding something
like `warehouse/static/css/*.css` to `.gitignore`? How would that
interact with the `manifest.json`?