Denormalize the sitemap bucket that an item belongs to
When many items exist in the database, pulling all the items into
the application and computing the buckets can take a long time.
Instead we'll denormalize this into the database so we can query
against it directly. We do this instead of dynamically computing
the bucket in the database because that still takes a significant
amount of time.