Switch from uWSGI to Gunicorn
While uWSGI has a lot of options and is theoretically better than
gunicorn, we're going to switch back to gunicorn. uWSGI breaks a lot
of the rules of a Python interpreter in order to squeeze out more
performance and the sum of all of these made working with it
difficult.

We'll use gunicorn with the gevent worker so that we don't block
threads when there are large uploads happening.