Speed up the index page by denormalizing counts
The SELECT COUNT(*) query in PostgreSQL is fairly slow so we'll work
around that by making a special table to record the row counts in
and use triggers to keep it updated on INSERTs and DELETEs. We can
then query against that table instead.

This also allows us to get the row counts for multiple tables in a
single query instead of in multiple queries.