Smartly compress our HTTP responses
Because of the CRIME attack we cannot just blindly compress our
content since that may lead to leakage of secure information. To
compress what we can we'll use a tween that checks for Vary headers
for Cookie and Authorization, if neither of those exist then there
is no sensitive information and we can safely compress.