Refactor warehouse.csrf to rely on view derivers
This greatly simplifies our code here since we can largely just reuse the built
in CSRF support in Pyramid now. We still need to write our own csrf view
deriver because our logic is a bit more careful about not making the outcome
of a particular view dependent on the contents of a Cookie.

In addition, we also drop the concept of "Scoped" CSRF tokens. While this
provides an extra assurance in the form that a single leaked token cannot be
used to attack other parts of the site, in practice it doesn't really add any
value because the additional origin check should, in theory, be good enough on
it's own when you're strictly a HTTPS only site like we are. In addition,
dropping this allows us to utilize more of the CSRF code from Pyramid itself
instead of reinventing more of it.