Prevent search pagination deeper than 10,000 items
Elasticsearch (and distributed search engines in general) get a pathological
behavior when attempting to paginate too deeply into their result set. Since
it's unlikely that anyone is going to need a result past the 500th page, we'll
just limit the total number of items that we can return to 10,000.

In addition, we'll make any attempt to view a page past the end of the result
set for a search query result in a 404 error.