Wrap "expected" xmlrpc errors to prevent logging
There are two kinds of exceptions that may occur in an XMLRPC view, ones that
are the fault of bad logic in the view, and ones that are the fault of the
remote client calling the view with the wrong inputs.

For the first of these, we want to raise the normal exceptions and have
pyramid_rpc log the fact they occurred. However, we don't want this to happen
with the other kind since we're only using them as a way to signal to the end
user they need to change something.

To handle this, we've created a subclass of xmlrpc.server.Fault which will wrap
a real exception and automatically translate this to an xmlrpc error. This will
avoid hitting the branch in pyramid_rpc that causes this exception to get
logged but will still signal an error back to the end user.