Remove support for localization with L20n
The L20n support was never really tested, never used, and has no champions able
to work on it currently. Changes are landing that are slowly causing what was
there to bit rot and it's in a weird semi-used state where we sometimes care
and sometimes don't, making everything harder and less consistent.

Removing this will make things easier to work on and more consistent, and if
someone comes forward to dedicate time to keeping Warehouse's translation
tooling working, we can add it back.