implement limits on response size for changelog and changelog_since_serial XMLRPC endpoints
This matches behavior introduced to XMLRPC in the pypi-legacy codebase.

I ended up tracking a reoccuring severe memory consumption issue to these methods.

When requested with an argument that is wildly behind of the current serial or date, these methods led to very large responses which were slow and cost us precious memory on the backend servers.