Disable Conditional GETs on file URLs (#2058)
Currently Fastly does not provide a way to access response headers when
the response is a 304 response. This is because the RFC states that only
a limited set of headers should be sent with a 304 response, and the
rest are SHOULD NOT. Since this stripping happens *prior* to vcl_deliver
being ran, that breaks our ability to log on 304 responses. Ideally at
some point Fastly offers us a way to access the "real" response headers
even for a 304 response, but for now, we are going to remove the headers
that allow a conditional response to be made. If at some point Fastly
does allow this, then we can revert this.