Stop using inline styles for cube SVGs
The Content Security Policy HTTP header prevents unsafe-inline
styles from executing, which means that in some browsers the cubes
might appear as black paths.

It's also possible to style these from a CSS/SCSS file instead, but
that might make the SVG files harder to edit in an external editor
in the future.

The styles are all valid SVG path attributes.