Add Sass linting
This linting uses the Airbnb `.scss-lint.yml` converted to the
sass-lint format.

https://github.com/airbnb/css/blob/9e3c56db12f96a5b2/.scss-lint.yml

Every rule is set to error (severity 2) except for:

- brace-style
- quotes

These both have many violations which I think it would be better to
clean up over time.

That's why I've set `--max-warnings 98` in the Makefile - this will
stop the number of warnings increasing but allow it to gradually
improve.