resolves journal display issue from #2377
@clgarciga and I investigated #2377 and determined that there was a
mismatch in the journals_query for the admin.project.journals view.

JournalEntry is stored with the non-normalized name for the project, and
the query previously was comparing Project.normalized_name to JournalEntry.name

one thing to note, I'm not sure if journals are updated when projects have been
renamed in the past. It might be worth adding a normalized_name to JournalEntry
for safety, or add uuids as @dstufft alluded in the past

Co-authored-by: <cglanza@gmail.com>