Make classifiers 'sticky' on /search/ results page
Change the template for /search/ so that the search form uses the
currently active classifiers also for the next search, instead of
clearing the classifier set.

The search forms on other pages are not changed, so the user can reset
the search by e.g. returning to the main page --- or, of course,
removing the classifiers one by one.