Move referrer security to HTTP header
Latest versions of Chrome and Firefox don't respect this part of
Content-Security-Policy:

Chrome:

> Unrecognized Content-Security-Policy directive 'referrer'.

Firefox:

> Content Security Policy: Referrer Directive ‘origin-when-cross-origin’ has been deprecated.
> Please use the Referrer-Policy header instead.

This commit uses the Referrer-Policy header as a new tween.

This is my first time making a change of this complexity so I've copied
the minimum amount of code possible from the CSP functionality to make
it work.