Password Reset (#2764)
* Add get_user_by_username utility

* Add PasswordRecoveryService

* Add a RecoverPasswordForm

* Add recover_password view

* Add ResetPasswordForm

* Add reset_password view

* Add 'Forgot Password' link to login page

* Don't use the term 'recovery'

We never actually are recovering a password: we're requesting a password
reset instead. This makes the difference between requesting a password
reset and actually resetting a password more clear.

* Use correct MAIL_SENDER

* Use templates for emails

This allows us to take much of the message content out of the view into
a template that is more consistently formatted.

* Move InvalidPasswordResetToken into .interfaces

So it can live alongside the TooManyFailedLogins exception

* Make a UserService fixture

This can be used as a fixture to the tests instead of constructing a
DatabaseUserService manually in the test body

* Make PasswordResetService part of the UserService

This doesn't really need to be a separate service, and things are
simpler if it's just part of the UserService, especially considering
that creating a PasswordResetService requires a UserService on
initialization anyways.

* Use request.params instead

This way we won't need to include the otk value in the form as it will
be part of the URL parameters

* Don't fail validation with existing password

There isn't really much advantage to doing this besides annoying the
user. If the passwords are in fact the same, the token will still be
expired and the password hash will be updated regardless

* Handle the case where the token might be missing

This is possible if a user visits the reset route without having the
parameter in the URL

* Flash on invalid token instead

If we've gotten an invalid token, just flash a message instead and
redirect to the password reset request view. Most likely, the token has
expired and the user needs to request a new one.

Also simplify the return values a bit.

* Change validate_otk to get_user_by_otk

And have it return a User object, not just an ID.

* Refactor account forms and mixins

Provide more discrete mixins for use with the various account related
forms to reduce duplication and make the forms more composable.

Also fixes an issue where the same user input fields being provided to
the PasswordStrengthValidator for Registration weren't being included
when resetting a password as well.

* Clean up and revise comments

* Assert on the call recorders

Also cleans up a few tests to use the user_service fixture

* Style reset password pages

* Add label to confirm password field(s)

* Make log in, log out consistent

* Flash on successful reset/login

* Simplify get_user_by_username

* Make token_max_age configurable

* Move tokens into service, better invalid messages

Pulls the token stuff out of the DatabaseUserService into its own
service.

Also ditches the hashing in the token in order to provide better
messages when the token is invalid.

* Tell the user how long the token is valid

In both the email and in the message after the email is sent.

* Inform users they can reply to the reset email

* Update reset info copywriting

* Increase clarity and helpfulness of `InvalidPasswordResetToken` messages

Since the exception message is flashed for the user, this is an OK place to add this.

It may be better to create subclasses of `InvalidPasswordResetToken` for each case so that messaging can be applied in the view, where translation could occur.

* Revise message when password has been changed