Role Management (#2705)
* Add 'manage' permission

* Always store principals as strings

Before, these were always UUID objects, but since #1329 this is stored
as a string for session-based authentication only. To keep everything
consistent, always use strings over UUID objects.

* Split profile and project management

This gives the logged in user a place to manage their profile, and a
place to manage their projects. Mostly stubbed out for now.

* Put the gravatar link on the 'Manage Profile' page

This no longer needs to be a client-side include because we can just
edit it via profile management when the user is logged in.

* Some really rudimentary styling, please revert

* Role management

Adding and deleting roles

* Update logged in information architecture, begin styling

* Make collaborator form more simple

* Allow stacking flash messages

* Reuse dropdown SCSS

* Style releases table

* Update management UI based on meeting feedback

* Fix linter errors

* Hide 'duplicate' roles

* Break management forms into mixins

* Simplify delete view

* Add ability to change existing roles

* Add JournalEntries when adding/removing roles

* Add labels for screen readers

* Add delete modals, tabs, clean up UI

* Put TODOs in comments

* Namespace manage routes

* Hide draft UIs in templates

* Properly comment out Edit Project link

* Change 'Preview' to 'View'

This will always link to a project/release that is live, so it's never
really a "Preview" per se. Also, this allows us to actually have a
"Preview" some day when we allow for staged releases.

* Fix more linting errors