Fix humanized times (#2924)
* Don't run timeAgo every minute

Not only was it taking 60 seconds for the humanized dates to appear, but
every date on the page was being re-humanized every 60 seconds.

* Fix casing on humanized date

* Don't use humanized dates for project history

It's more useful to see the actual timestamps here

* Use macro for consistent humanization

We had lost the trailing `Z` on the timestamp somehow which ensures that
the relative time takes the users timezone into account. This macro will
ensure humanized times are consistent site-wide.