Defer the Release.description column (#3045)
The vast bulk of the time, when querying against a Release we do not
want the description column, because the column is very large and can
easily spike memory and make queries take longer, particularly when
loading multiple Releases.

Instead we will defer the column so that fetching it requires making
another query or explicitly un-defering it. This will cause an extra
query to happen on views that do not have direct control over the
query (e.g. ones using the Pyramid Context objects) but the impact of
a single extra query on these pages is pretty low while the impact of
loading this column by default is playing a whack-a-mole game with any
query against Release to add a load_only() or defer() option.