better match legacy JSON documents (#3213)
There is one compromise here in that rather than _not_ sending a key for
`.info.requires_dist` we send a `null`. This is semantically the same, as we
can somewhat safely assume folks are calling the equivelant of `get` on the
object since the key may or may not exist.