add project.last_serial to simple.detail and legacy.api.json routes
This adds the project's latest serial to `/simple/<project_name>/` and `/pypi/<project_name>/json`.

As Warehouse uses Fastly Soft Purges, we need to tweak the ETag in order to notify the CDN that the page has changed. Currently the CDN performs a conditional GET in order to decide if it needs to pull refreshed content, and since the contents of these pages don't change for maintainer and role updates it skips purging them.

This is primarily a concern for clients like bandersnatch which use the X-PyPI-Last-Serial header to determine if they're getting the latest goods.