Remove roadmap link from footer
Since an out-of-date roadmap may be worse than none, and since we
predict that our updates and people's interest will diminish after
launch, we are removing the pypi.org link to the roadmap after the
Launch milestone. We'll retain links in other bits of our repo and our
developer docs, where people who look for it can will more easily see
if the roadmap or pages linking to it are stale.

Fixes #2851.