Gracefully handle redis errors while rate limiting
Instead of hard failing when there is a redis issue while rate limiting,
we will instead just assume that the rate limit is *not* being hit, and
allow the user to do what they were trying to do.

We will submit a metric anytime we have an error in the rate limiter.