Check request.unauthenticated_userid first
Sometimes we send emails in a context where we have an unauthenticated
user. Invoking request.user.id in this context (to get an authenticated
user ID) causes recursion in our auth mechanisms and ultimately failure.

Instead, we should first check whether there is an unauthenticated user,
and use that user ID for IP redaction before trying to use an
authenticated user's ID.