Ensure that headers returned from upload endpoint are valid ISO-8895-1
Because the legacy upload endpoint returns user supplied text, we need to guard against characters in the Status header that the web server (gunicorn in our case) cannot handle.

This roundtrips the message that may contain user supplied text through ISO-8895-1 encoding/decoding before sending it off as a header.