test: enforce unique generation of filenames
When generating a sufficiently large set of filenames, we saw collisions
in the database persistence, as they are expected to be unique.

The factory.Faker() interface doesn't currently provide us a hook to use
the underlying `faker.unique...` interface, so we have to use it
ourselves for now. A comment in the code links to the current open
issue.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>