Improvements to recovery codes (#10707)
* Add timestamp to recovery codes filename

* Validate password with require_reauth instead

* Don't require 2FA before recovery codes

* Refactor 2FA section

* Make button line-height the same as anchor

* Fix button formatting

* Continue to 2FA-only page instead

* Don't need to pass ip_address around

* Refactor recovery code logic for try/except flow

This makes it easier to return more detailed errors than 'false' from
the logic here. Raising exceptions also makes it more likely that we
won't forget to act on a return value in a failure case here.

* Add RecoveryCode.burned timestamp

* Mark recovery codes as burned instead of deleting

This makes it easier for us to tell when a user has burned a code.

* Fix tests

* Include username in recovery codes filename

* Send emails on recovery code use and regeneration

* Allow configurable reauth period

* Remove unused import

* Fix an issue with regeneration

Mixing 'require_reauth' and POST requests doesn't work. Since code
generation was already a GET request, and because this will always be
behind a 10s reauthentication, switch this to a GET request.

* Update translations