test: assert value on the exception location
The inner assertion was getting "swallowed" during pytest's assertion
rewriting, so the target location wasn't being validated.

Moving the assert outside of the context manager provides us access to
assert against the value.

Outdenting this assertion provided the visibility that the assertion was
incorrect, as the err.value is a Pyramid exception, which has a
`location` property we are validating against.

Refs: https://docs.pytest.org/en/6.2.x/assert.html#assertions-about-expected-exceptions
Refs: https://docs.pylonsproject.org/projects/pyramid/en/latest/api/httpexceptions.html#pyramid.httpexceptions.HTTPSeeOther

Signed-off-by: Mike Fiedler <miketheman@gmail.com>