chore(dev): set web to depend on db
When running `make tests`, the current expectation is that the user has
executed `make serve` and the database container is running in already.

This brings in all the other runtime dependencies, when all the tests
need is a running database instance.

Setting the `depends_on` flag tells docker-compose to start the db
container first, any time we run a `web` container, and saves us the
step of starting it ourselves in another shell.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>