chore(dev): set elasticsearch to single-node
The current settings for the elasticsearch container place the started
node in "production, single node cluster" mode.
This performs boostrap tests, one of which tests for virtual memory via
the `sysctl` setting of `vm.max_map_count`.

In certain Docker-in-Docker development environments, the end user may
not have access to the underlying host settings, and cannot modify this
setting.
An alternative is to set the `discovery-type` to a single node, placing
the service in "development mode", turning any bootstrap checks from
errors to warnings.

Refs: https://www.elastic.co/guide/en/elasticsearch/reference/7.10/system-config.html#dev-vs-prod
Refs: https://www.elastic.co/guide/en/elasticsearch/reference/7.10/bootstrap-checks.html

Originally reported in #1511, related to #4370

Signed-off-by: Mike Fiedler <miketheman@gmail.com>