correct logic in github-actions-deps
```
GITHUB_BASE_REF := $(shell echo "$${GITHUB_BASE_REF:-false}")

ifneq ($(GITHUB_BASE_REF), false)
...
```

in the Makefile, this set the value of the Make var to `false` if no GITHUB_BASE_REF existed in environment

Now we soley rely on [GitHub Actions env vars](https://docs.github.com/en/actions/learn-github-actions/environment-variables#default-environment-variables) since the Make variable doesn't propogate to the script.

"The name of the base ref or target branch of the pull request in a workflow run. This is only set when the event that triggers a workflow run is either pull_request or pull_request_target. For example, main."

So, on pushes to main this is empty.

Thus our current condtion:

```
if [[ -z "${GITHUB_BASE_REF}" ]]; then
```

Is actually the opposite effect of what we had previously.debug