feat(dev): express dependencies better
Currently, a developer is expected to perform `make serve` prior to
running `make initdb`, as the `serve` command will start **all**
services expressed in `docker-compose.yml`.

By using a `healthcheck` on `db` that uses a built-in CLI tool to test
for health, and Docker Compose 3.9 Long Syntax for the conditions, we
can now run `make initdb` in one command, followed by `make serve`.

This enables pre-build behaviors to establish the `postgres` and
`elasticsearch` databases prior to other apps starting.

Refs: https://docs.docker.com/compose/compose-file/#healthcheck
Refs: https://www.postgresql.org/docs/10/app-pg-isready.html
Refs: https://docs.docker.com/compose/compose-file/#long-syntax-1

Signed-off-by: Mike Fiedler <miketheman@gmail.com>