fix(dev): correct healtcheck command
While the previous command works, and reports that the db is indeed
ready and live for traffic, it emits wanrings in the logs that the
role named "root" does not exist, since it's executing under the user
context it was started with.
The healthcheck runs at 30 second intervals, so gets noisy.

Pass the explicit user role to prevent these warnings.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>