chore(deps): update all babel js deps
Update the babel packages that are used to their latest versions.
None are major version bumps, but some will now show deprecation
messages on installation, which should be handled at a later date.

Excluded is `babel-eslint` as it's been moved to `@babel/eslint-parser`
so will update that and any configuration associated when updating
`eslint`.

Resolves 9 npm vulnerabilities.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>