canonically serve JSON urls via normalized project names
something like 25 requests per second to the PyPI backends are *just* serving redirects from one name to another.

paired with forthcoming changes to our CDN configuration, we will begin serving JSON responses for any project name from the canonical URL that the project name normalizes to.

then the next step will be to announce this potentially breaking change and set a date at which we start serving the redirect from the edge.