Manage organization "Projects" (#11473)
* Remove is_active flag from Org Project model

* Add Organization Project services

* Add tests for Organization Project services

* Remove is_active flag from Org Project model...
I really mean it this time.

* Manage organization "Projects" with projects list

* "Add project" form in organization "Projects"

* "Transfer project" form in project "Settings"

* "Remove project" form in project "Settings"

* Connect "Add project" form to services

* Connect "Transfer project" form to services

* Connect "Remove project" form to services

* NFC: Refactor `add_project` as a function in utils

Moving the code from warehouse/forklift/legacy.py to a separate function
`warehouse.utils.project.add_project` will let us reuse it elsewhere.

* NFC: Rename {Create => Add}OrganizationProjectForm

- {CreateOrganizationProjectForm => AddOrganizationProjectForm}
- {create_organization_project_form => add_organization_project_form}

* "Add project" form with create new project option

* Disable <select> when there are no projects to add

* Email templates for organization projects

- "organization-project-added"
- "organization-project-removed"

* `send_*_email` functions for organization projects

- `send_organization_project_added`
- `send_organization_project_removed`

* NFC: Refactor `TestAddOrganizationProjectForm`

`@pytest.mark.parametrize` to reuse the same `test_validate` code.

* Send emails for organization projects

* Clearer labels in "Add project to organization" form

* Tighten permissions for "manage:organization"

* Require another owner to remove project from org

Project must have at least one individual owner before it can be
removed from an organization.

* Check for org that project can be transferred to

Don't show the "Transfer project" button if the user odes not own any
organizations that the project can be transferred to.

* Grant project permissions to organization owners

* Refactor `add_project` to avoid creating Owner

- Move Owner role creation out of the function
- Move validation code to separate `validate_project_name` function

* List projects owned through organizations

As @ewdurbin noted when reviewing #11473, I had not adjusted the queries
for a user's projects to account for projects managed by organizations
owned by the user. Fixed by adding `UNION` clauses to those queries.

* Add badge for projects owned by organizations

* Show organization projects in main projects list

Projects in organizations owned by the user will be displayed in the
user's main projects list.

* Update Alembic migration tree after merge 'main'

* Improve query for "sole owner"

@ewdurbin pointed out that the previous version of the query would fail
if a project with one individual owner is also owned by an organization.

* NFC: Project.{organizations[0] => organization}

Refactor relationship via association table to be one-to-many to avoid
confusion in the future.

* Remove user owner when transferring project to org

When a user transfers a project to an organization, remove that user as
an individual owner of the project. As an owner of the organization, the
user will still have full permissions over the project.

* Notification on `add_organization_project` success

Co-authored-by: sterbo <matt.sterba@gmail.com>
Co-authored-by: Ee Durbin <ewdurbin@gmail.com>