Set trial afterwards so checkout page looks better
Setting trial beforehand results in a checkout page that has "XX days
free" in large bold letters at the top. This could be confusing for
users because there is no place on the page for us to explain that the
purpose of the trial is to set the billing date to the first day of the
following month.