feat: support markdown style math tags
We can configure MathJax to detect the convention of single dollar signs
in addition to its other methods, as it's not on by default.

The inline script needs to be supported by our CSP, and if the stanza
ever changes, we'll need to recalculate the hash and update the CSP
accordingly.

Resolves #12009
Resolves pypa/readme_renderer#214

Signed-off-by: Mike Fiedler <miketheman@gmail.com>