chore: make insecure algorithm use explicit
When analyzing the codebase with security focused tools, usage of SHA-1
and MD5 algorithms trigger warnings that these algos should not be used
in a secure context, which they are not.

Since scanners have no way of knowing, use the `usedforsecurity` boolean
introduced in Python 3.9 to signal that these are not for security
purposes.

They are used for either filename hashing, or required conformance to an
external API structure.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>