Add admin include to 404 page for project pages (#12704)
* Determine project name during HTTPNotFound

There's two cases where we might know the project name that was
requested: if the project exists, and traversal has happened, but the
project has no releases, and if the project doesn't exist. This handles
both cases.

* Don't traverse project for the admin include

Instead, we run the query for the project inside the view. This allows
us to retain the project name and pass it to the underlying template.

* Determine if the project name has been prohibited

And pass the result to the underlying template for the admin panel

* Make admin include handle nonexistent project

We want to display this on 404 pages when there isn't a corresponding
project. This also handles whether the project name has been prohibited
or not.

* Include admin panel on error page if project_name

We only want to add the include on error pages if there's a
project_name, otherwise this does nothing.

* Update translations

* Update tests