Migrate frontend tooling to Webpack (#12728)
* chore: set package to private to prevent publishing

We don't ever expect to publish this set of code to an external
registry, so prevent that from even potentially happening.

Refs: https://docs.npmjs.com/cli/v9/configuring-npm/package-json#private

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): remove all gulp* dependencies

As a step towards replacement with webpack, remove all currently
installed gulp dependencies.

Does not account for dependencies installed for usage in the Gulpfile,
those will be handled in another commit.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): remove cssnano

Only used in Gulpfile. May return partially as part of
`css-minimizer-webpack-plugin` in the future.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): remove del package

Only used in Gulpfile. Likely to be replaced with a webpack built-in.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): remove vinyl-named package

Only used in Gulpfile. Likely to be replaced with a webpack built-in.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): remove zopflipng-bin package

Was included to support `gulp-image`, which is removed.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): remove webpack-stream package

It's a gulp plugin, and we don't need it anymore.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): move webpack to devDependencies

Since it's not a runtime dependency, move it to the dev section.

Add `webpack-cli` as well.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): add eslint-config-webpack plugin

So when we lint the configuration file, we don't trigger linter errors.

Refs: https://github.com/webpack-contrib/eslint-config-webpack

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: add npm script to run build command

Eventually will replace entries in `bin/*` scripts using the hard-coded
`node_modules` path.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: update build to account for webpack changes

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* test: update script runners to account for changes

- The build command is now an `npm run` command
- Webpack's sourcemaps changes the prefix for the output

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: add initial webpack configuration

By no means extensive, but a starting point to build upon.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: add compression plugin with basic config

Implements the same that Gulpfile did with both gzip and brotli
configurations.

Left a TODO in place to account for specific Brotli methods when we add
those inputs.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: add basic manifest plugin

Generally works. Need to evaluate once subdirectories are included.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): install plugins related to css/images

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: add webpack config for css/images

Image compression isn't handled yet.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* refactor: extract zxcvbn from the main bundle

Handle the vendor bundle as its own copy, so that we don't embed it into
the main `warehouse.js` bundle.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: compress images in production pipeline

As most of our images are not directly imported, we have previously used
the `CopyPlugin` to get them to where they need to be.
Adding a `minimizer.generator` will minify them based on the lossy
plugins (similar to `gulp-image` previously).

Only invoked when `mode: "production"` to save development cycles.

https://webpack.js.org/plugins/image-minimizer-webpack-plugin/

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(deps): add rtlcss handling packages

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: add webpack config for rtlcss

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(admin): add missing fontawesome svg file

Ad admin doesn't use a package, rather a pre-minified version,
compilation fails if a referenced asset is missing. We've lived with
this missing for admin for a long time, so this should have no negative
impact.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore(admin): add admin config to webpack

Since we have two widely different UI stacks today, refactor the config
to export two configuration objects, so that they will both be run with
the same command.

To reduce some of the code duplication, the plugins that do not differ
between configurations have been extracted for reuse.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: pass NODE_ENV to webpack for build-time

Setting `production` enables asset minimizations, which we want in
production, but not as much during development.

We can set an override in development via `docker-compose.override.yaml`
like so:

    static:
      environment:
        NODE_ENV: "production"

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* docs: update development documentation for webpack

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* chore: remove gulpfile

Now that it's all been handled, no need for this to stick around.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* lint: remove added eslint packages

The webpack plugin is intended to be used inline with webpack
configuration, and conflicts with our own `eslint:recommended` rules.

Effectively reverts d477e7d1e026fb8716c5551cca74dc74264ff5bf

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* lint(sass): add unit to color.mix($weight) property

Using `$weight` without a percentage emits a deprecation warning.

No behavior changes as a result.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

Signed-off-by: Mike Fiedler <miketheman@gmail.com>
Co-authored-by: Dustin Ingram <di@users.noreply.github.com>